requirements.txt
# Document and Excel handling
python-docx==1.2.0
pandas==3.0.0
openpyxl==3.1.5
numpy==2.4.1
python-dateutil==2.9.0
six==1.17.0
lxml==6.0.2
typing_extensions==4.9.0
requests==2.32.5

# GUI calendar widget for Tkinter
tkcalendar==1.6.1

# Document templating
docxtpl==0.20.2
